requirements.txt
Django==4.2.13
djangorestframework==3.14.0
django-cors-headers==4.3.1
authlib==1.3.0
auth0-python==5.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.32.3
gunicorn==21.2.0
